CHATBOT-PY/src/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
langchain>=0.1.5
langchain-community>=0.0.15
langchain-openai>=0.0.5
openai>=1.10.0
chromadb>=0.4.22
pydantic>=2.5.0
tiktoken>=0.5.0
